+++ src/main/java/com/plivo/api/models/media/MediaResponse.java
@@ -1,0 +1,7 @@
+package com.plivo.api.models.media;
+
+import com.plivo.api.models.base.BaseResponse;
+
+public class MediaResponse  extends BaseResponse {
+}
+