+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Change Log
 
-## [4.5.0](https://github.com/plivo/plivo-java/tree/v4.5.0) (2020-04-29)
+
+## [4.6.0](https://github.com/plivo/plivo-java/tree/v4.6.0) (2020-04-29)
 - Add V3 signature helper functions.
 
-## [4.4.5](https://github.com/plivo/plivo-java/tree/v4.4.5) (2020-01-06)
-- Add MMS and Powerpack support
+## [4.5.0](https://github.com/plivo/plivo-java/tree/v4.5.0) (2020-01-29)
+- Add application cascade delete support.
+- Add post call quality feedback API support.
+- Add Tollfree support for Powerpack
+- Add Media support.
+- Add MMS and Powerpack support.
 
 ## [4.4.4](https://github.com/plivo/plivo-java/tree/v4.4.4) (2019-11-19)
 - Add GetInput XML support